Specify Bash as the shell to be used in Makefile
At least one of the rules uses pattern matching which is specific
to Bash and not portable, e.g.

rm -f $(book).{pdf,epub,mobi}

GNU Make defaults to /bin/sh, which is not guaranteed to be Bash.
For example, on Ubuntu systems /bin/sh is a symlink to the Dash
shell, which does not support the file pattern above.

Specifying the shell to be used gets around this problem,
although this does of course assume that Bash is installed.